Add async, refactory delete work

+++ src/our-work/our-work.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, UseGuards } from '@nestjs/common';
+import { Injectable, NotFoundException, UseGuards } from '@nestjs/common';
 import { PrismaService } from 'src/prisma.service';
 import { OurWork } from './our-work.model';
 import { OurWorkDto } from './dto/create-our-work.dto';
@@ -16,18 +16,22 @@ export class OurWorkService {
         return this.prisma.ourWork.findMany({ where: { website } });
     }
 
-    findOne(id: number, website): Promise<OurWork> {
+    async findOne(id: number, website): Promise<OurWork> {
         return this.prisma.ourWork.findUnique({ where: { id, website } });
     }
 
-    update(id: number, OurWorkDto: OurWorkDto, website): Promise<OurWork> {
+    async update(id: number, OurWorkDto: OurWorkDto, website): Promise<OurWork> {
         return this.prisma.ourWork.update({
             where: { id, website },
             data: OurWorkDto
         });
     }
 
-    remove(id: number, website): Promise<OurWork> {
+    async remove(id: number, website): Promise<OurWork> {
+        const work = await this.prisma.ourService.findUnique({ where: { id, website } });
+        if (!work) {
+            throw new NotFoundException('Work not found.');
+        }
         return this.prisma.ourWork.delete({ where: { id, website } });
     }
 }
